backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
aiosqlite>=0.20.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.26.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
tenacity>=8.2.0
redis>=5.0.0

# 测试依赖
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
fakeredis>=2.18.0
coverage>=7.3.0
